requirements.txt
aiogram==3.4.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
python-dotenv==1.0.1
dateparser==1.2.0
pydantic==2.4.2
pydantic-settings==2.2.1
httpx==0.27.0
openai==1.51.0
google-generativeai==0.8.3
